# modelrag-server/src/main/resources/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ModelRAG API v2",
    "version": "1.0.0",
    "description": "Typed, bearer-authenticated API for knowledge bases, RAG, conversations, memory, tools and evaluation."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/api/v2/assistant/streams": {
      "post": {
        "tags": ["assistant"],
        "operationId": "streamAssistant",
        "requestBody": { "$ref": "#/components/requestBodies/AssistantAsk" },
        "responses": { "200": { "$ref": "#/components/responses/SseStream" }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/v2/assistant/streams/{executionId}": {
      "get": {
        "tags": ["assistant"],
        "operationId": "streamAgentExecution",
        "parameters": [{ "$ref": "#/components/parameters/ExecutionId" }],
        "responses": { "200": { "$ref": "#/components/responses/SseStream" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/api/v2/assistant/executions/{executionId}": {
      "delete": {
        "tags": ["assistant"],
        "operationId": "cancelAssistantExecution",
        "parameters": [{ "$ref": "#/components/parameters/ExecutionId" }],
        "responses": { "200": { "$ref": "#/components/responses/EmptySuccess" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v2/assistant/auto": {
      "post": {
        "tags": ["assistant"],
        "operationId": "autoAssistant",
        "requestBody": { "$ref": "#/components/requestBodies/AutoQuestion" },
        "responses": { "200": { "$ref": "#/components/responses/AutoAnswer" }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/v2/assistant/auto/stream": {
      "post": {
        "tags": ["assistant"],
        "operationId": "streamAutoAssistant",
        "requestBody": { "$ref": "#/components/requestBodies/AutoQuestion" },
        "responses": { "200": { "$ref": "#/components/responses/SseStream" }, "400": { "$ref": "#/components/responses/ValidationError" } }
      }
    },
...
    "/api/v2/documents/{documentId}": {
      "delete": { "tags": ["documents"], "operationId": "deleteDocument", "parameters": [{ "$ref": "#/components/parameters/DocumentId" }], "responses": { "200": { "$ref": "#/components/responses/EmptySuccess" } } }
    },
    "/api/v2/datasets/{datasetId}/documents/{documentId}/chunks": {
      "get": { "tags": ["documents"], "operationId": "listDocumentChunks", "parameters": [{ "$ref": "#/components/parameters/DatasetId" }, { "$ref": "#/components/parameters/DocumentId" }], "responses": { "200": { "$ref": "#/components/responses/ChunkList" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "ModelRAG-HMAC" }
    },
    "parameters": {
      "DatasetId": { "name": "datasetId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "DocumentId": { "name": "documentId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "ExecutionId": { "name": "executionId", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } },
...
    },
    "requestBodies": {
      "AssistantAsk": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssistantAskRequest" } } } },
      "AutoQuestion": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AutoQuestionRequest" } } } },
...
    },
    "responses": {
...
      "AutoAnswer": { "description": "Auto-routed answer", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AutoQaResponse" } } } },
...
      "ChunkList": { "description": "Chunk list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } },
...
      "EmptySuccess": { "description": "Operation completed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } },
      "SseStream": { "description": "Server-sent events. Events include ACCEPTED, AGENT_STARTED, PLAN, ACT, OBSERVE, TOKEN, ANSWER, APPROVAL_REQUIRED, CANCELLED, DONE, HEARTBEAT or ERROR.", "content": { "text/event-stream": { "schema": { "type": "string" } } } },
      "ValidationError": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Unauthorized": { "description": "Missing or invalid bearer token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Forbidden": { "description": "User is not allowed to access the resource", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    },
    "schemas": {
      "ApiResponse": { "type": "object", "required": ["code", "msg", "timestamp"], "properties": { "code": { "type": "integer", "format": "int32" }, "msg": { "type": "string" }, "data": {}, "timestamp": { "type": "integer", "format": "int64" } } },
      "ErrorResponse": { "$ref": "#/components/schemas/ApiResponse" },
      "AssistantAskRequest": { "type": "object", "required": ["datasetId", "question"], "properties": { "datasetId": { "type": "integer", "format": "int64", "minimum": 1 }, "question": { "type": "string", "minLength": 1, "maxLength": 20000 }, "conversationId": { "type": ["integer", "null"], "format": "int64" }, "agent": { "type": "boolean", "default": false } } },
      "AutoQuestionRequest": { "type": "object", "required": ["question"], "properties": { "question": { "type": "string", "minLength": 1, "maxLength": 20000 }, "conversationId": { "type": ["integer", "null"], "format": "int64" } } },
...
      "AutoQaResponse": { "allOf": [{ "$ref": "#/components/schemas/ApiResponse" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/AutoQa" } } }] },
      "AutoQa": { "type": "object", "properties": { "datasetId": { "type": "integer", "format": "int64" }, "datasetName": { "type": "string" }, "route": { "type": "string" }, "status": { "type": "string" }, "answer": { "type": "string" }, "citations": { "type": "array", "items": { "$ref": "#/components/schemas/Citation" } }, "confidence": { "type": "number" }, "refused": { "type": "boolean" }, "traceId": { "type": ["string", "null"] }, "executionId": { "type": ["string", "null"] }, "approvalId": { "type": ["string", "null"] }, "steps": { "type": "array", "items": { "type": "string" } }, "degradedComponents": { "type": "array", "items": { "type": "string" } } } },
      "Citation": { "type": "object", "required": ["chunkId", "documentId", "documentName", "location", "indexVersion", "excerpt", "score"], "properties": { "chunkId": { "type": "integer", "format": "int64" }, "documentId": { "type": "integer", "format": "int64" }, "documentName": { "type": "string" }, "location": { "type": "string", "description": "PDF 页码或结构化章节位置" }, "indexVersion": { "type": "integer", "format": "int64" }, "excerpt": { "type": "string" }, "score": { "type": "number" } } },
...
    }
  }
}